+++ src/LiquidElectrolytes.jl
@@ -1,27 +1,34 @@
-module LiquidElectrolytes
-using VoronoiFVM, ExtendableGrids
-using DocStringExtensions
-using ProgressLogging
-using StaticArrays
-using LinearAlgebra
-using NLsolve
-using Unitful, LessUnitful
-import SciMLBase
+"""
+    LiquidElectrolytes
 
+$(read(joinpath(@__DIR__, "..", "README.md"), String))
+"""
+module LiquidElectrolytes
 using Base: @kwdef
+using DocStringExtensions: DocStringExtensions,  TYPEDEF, TYPEDFIELDS
+using ExtendableGrids: ExtendableGrids, num_nodes
+using LessUnitful: LessUnitful, @local_phconstants, @local_unitfactors, @ph_str, @ufac_str
+using InteractiveUtils: InteractiveUtils
+using Markdown: @md_str
+using Printf: Printf, @printf, @sprintf
+using ProgressMeter: ProgressMeter
+using ProgressLogging: @withprogress, @logprogress
+using SciMLBase: SciMLBase, solve!
+using VoronoiFVM: VoronoiFVM, TransientSolution, enable_boundary_species!, solve, testfunction, unknowns
+using VoronoiFVM: boundary_dirichlet!, fbernoulli_pm, SolverControl
+using LinearAlgebra: LinearAlgebra
 
 function __init__()
     return LessUnitful.ensureSIBase()
 end
 
-include("tools.jl")
+include("utils.jl")
 export RExp, RLog
 VERSION >= v"1.11.0-DEV.469" && eval(Meta.parse("public rexp, rlog"))
 
 include("electrolyte.jl")
 export ElectrolyteData, AbstractElectrolyteData
-export dlcap0, chargedensity, chemical_potentials!, rrate, debyelength, chemical_potential, c0_barc
-export showstruct, electrolyte, solventconcentration
+export dlcap0, chargedensity, chemical_potentials!, rrate, debyelength, chemical_potential, c0_barc, solventconcentration
 export isincompressible, iselectroneutral
 export chemical_potentials, electrochemical_potentials
 
@@ -35,15 +42,14 @@ include("pbsystem.jl")
 export PBSystem
 
 
-include("cells.jl")
-export ivsweep, dlcapsweep, currents, voltages_solutions, voltages_dlcaps, voltages_currents
-export AbstractSimulationResult, DLCapSweepResult, IVSweepResult
-
-include("equilibrium-pluto.jl")
-export EquilibriumData, create_equilibrium_system, create_equilibrium_pp_system,  dlcapsweep_equi
-VERSION >= v"1.11.0-DEV.469" && eval(Meta.parse("public set_molarity!, apply_voltage!, update_derived!"))
-VERSION >= v"1.11.0-DEV.469" && eval(Meta.parse("public calc_cmol, calc_QBL, calc_φ, calc_p, ysum"))
-include("equilibrium-supplement.jl")
+include("results.jl")
+export AbstractSimulationResult, voltages, currents, voltages_solutions, voltages_dlcaps, voltages_currents
+include("dlcapsweep.jl")
+export dlcapsweep,  DLCapSweepResult
+include("ivsweep.jl")
+export ivsweep, IVSweepResult
+include("cvsweep.jl")
+export SawTooth, cvsweep, CVSweepResult, period
 
 include("pnpstokes.jl")
 export PNPStokesSolver