+++ test/LiquidElectrolytesReTest.jl
@@ -1,0 +1,6 @@
+# See https://wias-pdelib.github.io/VoronoiFVM.jl/stable/devel/#ReTest
+
+module LiquidElectrolytesReTest
+using ReTest
+include("alltests.jl")
+end